# Conflicts:
#	src/main/java/org/bleachhack/gui/AccountManagerScreen.java
#	src/main/java/org/bleachhack/module/mods/AutoParkour.java
#	src/main/java/org/bleachhack/module/mods/Dispenser32k.java
#	src/main/java/org/bleachhack/module/mods/Nametags.java
#	src/main/java/org/bleachhack/module/mods/NewChunks.java
#	src/main/java/org/bleachhack/module/mods/Search.java

+++ src/main/java/org/bleachhack/util/render/Renderer.java
@@ -223,7 +223,7 @@ public class Renderer {
 			return Vec3d.ZERO;
 		}
 
-		double tickDelta = (double) MinecraftClient.getInstance().getTickDelta();
+		double tickDelta = MinecraftClient.getInstance().getTickDelta();
 		return new Vec3d(
 				e.getX() - MathHelper.lerp(tickDelta, e.lastRenderX, e.getX()),
 				e.getY() - MathHelper.lerp(tickDelta, e.lastRenderY, e.getY()),
